refactor: cameras3 to cameras/cameras2++
