+++ src/ros/cameras/cameras2++/include/cameras/cameras.hpp
