+++ src/ros/cameras/cameras2++/params/platform/aftershock_laptop/core.yaml
