+++ src/ros/cameras/cameras2++/params/platform/aftershock_laptop/payload/science-arc.yaml
