+++ src/ros/cameras/cameras2++/params/platform/intel_laptop/core.yaml
