+++ src/ros/cameras/cameras2++/params/platform/orin/core.yaml
