+++ src/ros/cameras/cameras2++/params/platform/orin/payload/ec.yaml
