+++ src/ros/cameras/cameras2++/params/platform/orin/payload/science-arc.yaml
