+++ src/ros/cameras/cameras2++/params/platform/orin_dev_board/core.yaml
