+++ src/ros/cameras/cameras2++/params/platform/rover/core.yaml
