+++ src/ros/cameras/cameras2++/params/platform/rover/payload/ec.yaml
