+++ src/ros/cameras/cameras2++/params/platform/rover/payload/science-arc.yaml
