+++ src/ros/cameras/cameras2++/params/platform/rover/payload/science_urc.yaml
