+++ src/ros/cameras/cameras2++/params/platform/tx2_dev_board/core.yaml
