+++ src/ros/cameras/cameras2++/src/directory_parameters.yaml
