+++ src/ros/cameras/cameras2++/src/streamer_parameters.yaml
