+++ src/ros/cameras/default.nix
@@ -1,0 +1,5 @@
+{
+  rosPackages = pkgs: with pkgs; {
+    nova-cameras = callPackage ./cameras2++ { };
+  };
+}
